pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "lifeoraai"
version = "0.1.0"
description = "Retrieval-Augmented Generation for health, nutrition, exercise, and lifestyle guidance"
requires-python = ">=3.11"
dependencies = [
    "sentence-transformers>=2.7.0",
    "chromadb>=0.5.0",
    "anthropic>=0.28.0",
    "ollama>=0.2.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "matplotlib>=3.8.0",
    "scikit-learn>=1.4.0",
]

[project.scripts]
lifeoraai = "main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["rag*", "retrieval*", "core*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
